Refactor files and types
Moved data extractors to a separate file.
Updated imports in all dependent modules.

This change also adds library level exports of all the major modules.
