+++ src/index.ts
@@ -1,1 +1,4 @@
 export * from './Auth0LoginProcessor'
+export * from './Types'
+export * from './DataExtractors'
+export * from './ExtendedError'
